Some simple tools for checking network interfaces
At times I find the router is in a bizarre state of:

WAN is down
LAN is up and router switching fine
Web server is down (so no LuCI)
ssh works

For these times I want a few simpel CLI tools to see what's going on I guess. And this is a start.