First working copy of the DDNS diagnostics hosted on a remote site
Because if the router experiences a WAN IP change for any reason (reboot say) and something goes wrong with the DDNS updates, and I am remotely siuated I have no way of knowing how to access my LAN. 

To wit, I now host a DDNS diagnostics page on a remote site with a fixed IP that I can access which has two key benefits: 1) I can always get to it and 2) it can use the NameCheap API becasue that API needs client IPs to be whitelisted and if the client IP changes voila can't use the API. 

So this is an experiment with WAN IP logging on the remote site. In this present setup I have it installed as a hotplug on ifup and a cron hourly, daily and on reboot for now, while I watch what works.