Fixed the WAN IP log to only send an update if a change in WAN IP is noticed
This would mean I can leave an hourly cron running indefinitely in case of issues as it will only send if the IP has chnaged. I would like to see the hotplug update working, but will wait until the next router restart to see if it did. No mood to change my WAN IP just for the test.